Rearrange tutorial content service tests
